+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/jacoco/csv/ConsoleCoverageReportOutputStream.kt
@@ -8,7 +8,7 @@ internal class ConsoleCoverageReportOutputStream(
     private val outputStream: OutputStream
 ) : OutputStream() {
 
-    private var byteArrayOutputStream = ByteArrayOutputStream()
+    private val byteArrayOutputStream = ByteArrayOutputStream()
 
     override fun write(b: Int) = byteArrayOutputStream.write(b)